오답 수정

+++ src/java_01/quiz1/Quiz01.java
@@ -2,6 +2,6 @@ package java_01.quiz1;
 
 public class Quiz01 {
     /*
-    long > float > double > int > short = char > byte
+    double > float > long > int > short = char > byte
      */
 }
